[Keymap] Fix build error `helix:five_rows` (#16847)
