+++ keyboards/helix/rev3_5rows/keymaps/five_rows/rules.mk
@@ -1,3 +1,5 @@
+USER_NAME := mtei
+
  CONSOLE_ENABLE = no        # Console for debug
  COMMAND_ENABLE = no        # Commands for debug and configuration
  # CONSOLE_ENABLE and COMMAND_ENABLE
